implement pipeline run

+++ pkg/test/runner/runner.go
@@ -23,8 +23,8 @@ const (
 	expectedResultsFile string = "results.yaml"
 	expectedDiffFile    string = "diff.patch"
 	expectedConfigFile  string = "config.yaml"
-	CommandFn           string = "fn"
-	CommandPipeline     string = "pipeline"
+	CommandFnEval       string = "run"
+	CommandFnRender     string = "render"
 )
 
 // NewRunner returns a new runner for pkg
@@ -45,10 +45,14 @@ func NewRunner(testCase TestCase, c string) (*Runner, error) {
 
 // Run runs the test.
 func (r *Runner) Run() error {
-	if r.cmd == CommandFn {
+	switch r.cmd {
+	case CommandFnEval:
 		return r.runFn()
+	case CommandFnRender:
+		return r.runPipeline()
+	default:
+		return fmt.Errorf("invalid command %s", r.cmd)
 	}
-	return r.runPipeline()
 }
 
 func (r *Runner) runFn() error {
@@ -79,7 +83,7 @@ func (r *Runner) runFn() error {
 	}
 
 	// run function
-	kptArgs := []string{r.cmd, "run", tmpPkgPath, "--results-dir", resultsPath}
+	kptArgs := []string{"fn", "run", tmpPkgPath, "--results-dir", resultsPath}
 	if r.testCase.Config.Network {
 		kptArgs = append(kptArgs, "--network")
 	}
@@ -114,7 +118,10 @@ func (r *Runner) runPipeline() error {
 	if err != nil {
 		return fmt.Errorf("failed to create temporary dir: %w", err)
 	}
-	defer os.RemoveAll(tmpDir)
+	if !r.testCase.Config.Debug {
+		// if debug is true, keep the test directory around for debugging
+		defer os.RemoveAll(tmpDir)
+	}
 	tmpPkgPath := filepath.Join(tmpDir, r.pkgName)
 	// create dir to store untouched pkg to compare against
 	orgPkgPath := filepath.Join(tmpDir, "original")
@@ -142,7 +149,7 @@ func (r *Runner) runPipeline() error {
 	// run function
 	var fnErr error
 	command := run.GetMain()
-	kptArgs := []string{r.cmd, "run", tmpPkgPath}
+	kptArgs := []string{"fn", "render", tmpPkgPath}
 	for i := 0; i < r.testCase.Config.RunCount; i++ {
 		command.SetArgs(kptArgs)
 		fnErr = command.Execute()
@@ -152,13 +159,6 @@ func (r *Runner) runPipeline() error {
 			}
 			break
 		}
-	}
-
-	// run formatter
-	command.SetArgs([]string{"cfg", "fmt", tmpPkgPath})
-	err = command.Execute()
-	if err != nil {
-		return fmt.Errorf("failed to run kpt cfg fmt: %w", err)
 	}
 
 	// compare results
